+++ version-info.js
@@ -1,0 +1,4 @@
+window.HrrraVersionInfo = Object.freeze({
+  versionCode: 7,
+  versionName: "1.0.6"
+});